+++ src/us/rynet/akaifire/receivers/ControlChangeReceiver.java
@@ -13,9 +13,9 @@ public class ControlChangeReceiver extends MidiReceiver {
   }
 
   public void send(MidiMessage message, long timeStamp) {
-    int status        = message.getStatus();
-    int controlNumber = (int)(message.getMessage()[1] & 0xFF);
-    int value         = (int)(message.getMessage()[2] & 0xFF);
+    byte status        = (byte)message.getStatus();
+    byte controlNumber = (byte)(message.getMessage()[1] & 0xFF);
+    byte value         = (byte)(message.getMessage()[2] & 0xFF);
 
     if (status == AkaiFire.CONTROL_START || status == AkaiFire.CONTROL_STOP) {
       akaiFire.getControls().stream().filter(control -> (control instanceof Button) && control.getIndex() == controlNumber).forEach(control -> ((Button)control).setPressed(status == AkaiFire.CONTROL_START)); // TODO: Should be ALL controls - rename setPressed to isChanging or something
